refactor: Extract output execution to its own component.
