+++ apps/dave/src/components/output/OutputView.tsx
@@ -1,57 +1,26 @@
 "use client";
-import type {
-    DelegateCallVoucher,
-    GetOutputReturnType,
-    Notice,
-    Output,
-    Voucher,
-} from "@cartesi/viem";
-import {
-    useEpoch,
-    useReadApplicationWasOutputExecuted,
-    useSimulateApplicationExecuteOutput,
-    useWriteApplicationExecuteOutput,
-} from "@cartesi/wagmi";
-import {
-    Alert,
-    Badge,
-    Button,
-    Divider,
-    Fieldset,
-    Group,
-    Spoiler,
-    Stack,
-    Text,
-    Tooltip,
-    useMantineTheme,
-} from "@mantine/core";
+import { Divider, Fieldset, Group, Spoiler, Text } from "@mantine/core";
 import { notifications } from "@mantine/notifications";
-import { useChainModal, useConnectModal } from "@rainbow-me/rainbowkit";
-import { useQueryClient } from "@tanstack/react-query";
-import { isNil, isNotEmpty, isNotNil, pathOr } from "ramda";
-import { isNotNilOrEmpty, isObj, isString } from "ramda-adjunct";
-import { Activity, useEffect, useMemo, type FC } from "react";
-import { TbExclamationCircle, TbInfoCircle, TbReceipt } from "react-icons/tb";
+import { isNotNil } from "ramda";
+import { type FC } from "react";
 import { formatUnits, isHex, type Hex } from "viem";
-import { useAccount, useWaitForTransactionReceipt } from "wagmi";
 import { useIsSmallDevice } from "../../hooks/useIsSmallDevice";
 import { getDecoder } from "../../lib/decoders";
 import Address from "../Address";
-import { useSelectedNodeConnection } from "../connection/hooks";
 import JSONViewer from "../JSONViewer";
 import useVoucherDecoder from "../specification/hooks/useVoucherDecoder";
-import TransactionHash from "../TransactionHash";
-import type { DecoderType } from "../types";
+import OutputExecution from "./OutputExecution";
+import type {
+    NoticeContentProps,
+    OutputViewProps,
+    VoucherContentProps,
+    VoucherOutput,
+} from "./types";
 
-interface OutputViewProps {
-    displayAs?: DecoderType;
-    output: GetOutputReturnType;
-    application: string;
-}
-
-type NoticeProps = { decodedData: Notice; decoderType: DecoderType };
-
-const NoticeContent: FC<NoticeProps> = ({ decodedData, decoderType }) => {
+const NoticeContent: FC<NoticeContentProps> = ({
+    decodedData,
+    decoderType,
+}) => {
     const decoderFn = getDecoder(decoderType);
 
     return (
@@ -65,28 +34,7 @@ const NoticeContent: FC<NoticeProps> = ({ decodedData, decoderType }) => {
     );
 };
 
-type VoucherOutput = Omit<Output, "decodedData"> & {
-    decodedData: Voucher | DelegateCallVoucher;
-};
-
-type VoucherProps = {
-    output: VoucherOutput;
-    application: string;
-    decoderType: DecoderType;
-    title?: string;
-};
-
-type Proof = { outputIndex: bigint; outputHashesSiblings: Hex[] };
-
-const buildProof = (output: VoucherOutput): Proof => ({
-    outputIndex: output.index,
-    outputHashesSiblings: output.outputHashesSiblings ?? [],
-});
-
-// xxx: Maybe should become an environment var.
-const POLLING_INTERVAL_MS = 8000 as const;
-
-const VoucherContent: FC<VoucherProps> = ({
+const VoucherContent: FC<VoucherContentProps> = ({
     decoderType,
     output,
     application,
@@ -94,16 +42,7 @@ const VoucherContent: FC<VoucherProps> = ({
 }) => {
     const decoderFn = getDecoder(decoderType);
     const { isSmallDevice } = useIsSmallDevice();
-    const userAccount = useAccount();
-    const connectModal = useConnectModal();
-    const selectedNode = useSelectedNodeConnection();
-    const chainModal = useChainModal();
-    const queryClient = useQueryClient();
-
-    const { decodedData, epochIndex } = output;
-
-    const theme = useMantineTheme();
-    const epochQuery = useEpoch({ epochIndex, application });
+    const { decodedData } = output;
     const hasPayload =
         isHex(decodedData.payload) && decodedData.payload !== "0x";
     const amount = decodedData.type === "Voucher" ? decodedData.value : 0n;
@@ -111,115 +50,6 @@ const VoucherContent: FC<VoucherProps> = ({
     const voucherDecodingRes = useVoucherDecoder({ voucher: decodedData });
     const hasDecodedData = isNotNil(voucherDecodingRes.data);
     const isDecodedSelected = decoderType === "decoded";
-    const appAddress = application as Hex;
-
-    const isClaimAccepted = epochQuery.data?.status === "CLAIM_ACCEPTED";
-    const hasExecutionTransaction = isNotNil(output.executionTransactionHash);
-
-    const wasOutputExecutedQuery = useReadApplicationWasOutputExecuted({
-        address: appAddress,
-        args: [output.index],
-        query: {
-            enabled: !hasExecutionTransaction && isNotNil(output.index),
-        },
-    });
-
-    const {
-        data: wasOutputExecuted,
-        isFetching: checkingOutputExecuted,
-        error: checkingOutputExecutedError,
-        refetch: recheckOutputExecuted,
-    } = wasOutputExecutedQuery;
-
-    const proof = buildProof(output);
-
-    const hasHashes = isNotEmpty(proof.outputHashesSiblings);
-    const canSimulate =
-        userAccount.isConnected &&
-        !checkingOutputExecuted &&
-        wasOutputExecuted === false &&
-        hasHashes &&
-        isClaimAccepted;
-
-    const prepare = useSimulateApplicationExecuteOutput({
-        address: appAddress,
-        args: [output.rawData, proof],
-        query: {
-            enabled: canSimulate,
-        },
-    });
-
-    const execute = useWriteApplicationExecuteOutput();
-    const wait = useWaitForTransactionReceipt({
-        hash: execute.data,
-    });
-
-    const isExecutingVoucher = execute.isPending || wait.isFetching;
-
-    const errors = useMemo(() => {
-        return [checkingOutputExecutedError, prepare.error].filter(
-            isNotNilOrEmpty,
-        );
-    }, [checkingOutputExecutedError, prepare.error]);
-
-    const hasErrors = errors.length > 0;
-
-    const isExecuteDisabled =
-        !isClaimAccepted ||
-        checkingOutputExecuted ||
-        wasOutputExecuted ||
-        prepare.isFetching ||
-        hasErrors;
-
-    useEffect(() => {
-        if (wait.isSuccess) {
-            notifications.show({
-                title: "Voucher execution status",
-                message: "Executed successfully",
-                color: "green",
-                withBorder: true,
-            });
-            execute.reset();
-            recheckOutputExecuted();
-            // mark outputs as stale to get latest values from rollups-node.
-            queryClient.invalidateQueries({ queryKey: ["outputs"] });
-        }
-    }, [wait.isSuccess, recheckOutputExecuted, execute, queryClient]);
-
-    useEffect(() => {
-        let key: unknown;
-        // skip invalidation if there is no node-connected or the connected node is a mock.
-        if (
-            isNotNil(selectedNode) &&
-            selectedNode.type !== "system_mock" &&
-            !isClaimAccepted
-        ) {
-            key = setInterval(() => {
-                // Invalidation makes the targeted queries stale.
-                // It creates a Polling effect.
-                queryClient.invalidateQueries({
-                    predicate: (query) => {
-                        const [baseKey, paramsKey] = query.queryKey;
-                        const isEpochQuery =
-                            isString(baseKey) && baseKey.includes("epoch");
-                        const isMatchingEpochIndex =
-                            isObj(paramsKey) &&
-                            pathOr("", ["epochIndex"], paramsKey) ===
-                                output.epochIndex.toString();
-
-                        return isEpochQuery && isMatchingEpochIndex;
-                    },
-                });
-            }, POLLING_INTERVAL_MS);
-        }
-
-        return () => {
-            if (isNotNil(key)) {
-                console.info(`Clearing the polling.`);
-                clearInterval(key as number);
-            }
-        };
-    }, [queryClient, isClaimAccepted, selectedNode, output.epochIndex]);
 
     return (
         <Fieldset legend={title}>
@@ -259,107 +89,19 @@ const VoucherContent: FC<VoucherProps> = ({
                     </Spoiler>
                 </>
             )}
-            <Stack>
-                <Divider mt="sm" />
 
-                <Activity mode={!isClaimAccepted ? "visible" : "hidden"}>
-                    <Group justify="flex-end">
-                        <Badge
-                            radius="xs"
-                            size="lg"
-                            leftSection={
-                                <Tooltip
-                                    label={
-                                        "Once the epoch status become CLAIM_ACCEPTED the voucher will become executable."
-                                    }
-                                >
-                                    <TbInfoCircle
-                                        size={theme.other.smIconSize}
-                                    />
-                                </Tooltip>
-                            }
-                        >
-                            Waiting Claim
-                        </Badge>
-                    </Group>
-                </Activity>
-
-                <Activity mode={isClaimAccepted ? "visible" : "hidden"}>
-                    <Group
-                        justify={
-                            hasExecutionTransaction
-                                ? "space-between"
-                                : "flex-end"
-                        }
-                    >
-                        {isNotNil(output.executionTransactionHash) && (
-                            <Group gap={3}>
-                                <Tooltip label="Transaction hash">
-                                    <TbReceipt size={theme.other.mdIconSize} />
-                                </Tooltip>
-                                <TransactionHash
-                                    transactionHash={
-                                        output.executionTransactionHash
-                                    }
-                                />
-                            </Group>
-                        )}
-                        <Button
-                            disabled={isExecuteDisabled}
-                            loading={isExecutingVoucher}
-                            onClick={() => {
-                                const needSwitchNetwork =
-                                    userAccount.isConnected &&
-                                    isNil(userAccount.chain);
-                                const canSend =
-                                    !needSwitchNetwork &&
-                                    userAccount.isConnected;
-
-                                if (needSwitchNetwork)
-                                    return chainModal.openChainModal?.();
-
-                                if (canSend) {
-                                    execute.writeContract(
-                                        prepare.data!.request,
-                                    );
-                                } else {
-                                    connectModal.openConnectModal?.();
-                                }
-                            }}
-                        >
-                            {wasOutputExecuted
-                                ? "Executed"
-                                : checkingOutputExecuted
-                                  ? "Checking voucher..."
-                                  : prepare.isFetching
-                                    ? "Preparing voucher..."
-                                    : "Execute"}
-                        </Button>
-                    </Group>
-                </Activity>
-            </Stack>
-            {hasErrors && (
-                <Stack py="sm">
-                    {errors.map((error) => (
-                        <Alert
-                            key={error?.message}
-                            color="red"
-                            title={pathOr("", ["shortMessage"], error)}
-                            icon={
-                                <TbExclamationCircle
-                                    size={theme.other.mdIconSize}
-                                />
-                            }
-                        >
-                            {pathOr(
-                                "Something went wrong!",
-                                ["cause", "data", "errorName"],
-                                error,
-                            )}
-                        </Alert>
-                    ))}
-                </Stack>
-            )}
+            <OutputExecution
+                application={application as Hex}
+                output={output}
+                onSuccess={() => {
+                    notifications.show({
+                        title: "Voucher execution status",
+                        message: "Executed successfully",
+                        color: "green",
+                        withBorder: true,
+                    });
+                }}
+            />
         </Fieldset>
     );
 };
